+++ test/runtests.jl
@@ -14,6 +14,7 @@ const TEST_MODE = get(ENV, "TEST_MODE", "default")
         include("unit/test_overrelaxation.jl")
         include("unit/test_pt.jl")
         include("unit/test_checkpoint.jl")
+        include("unit/test_geometry.jl")
     end
     if TEST_MODE in ("default", "all", "aqua")
         include("aqua.jl")